Split fixtures in the Plugin module
