+++ saleor/plugins/tests/fixtures/__init__.py
@@ -1,0 +1,3 @@
+from .gateway import *  # noqa: F403
+from .plugin_configuration import *  # noqa: F403
+from .plugin_manager import *  # noqa: F403
