+++ saleor/plugins/tests/fixtures/gateway.py
@@ -1,0 +1,18 @@
+import pytest
+
+
+@pytest.fixture(autouse=True)
+def setup_dummy_gateways(settings):
+    settings.PLUGINS = [
+        "saleor.payment.gateways.dummy.plugin.DummyGatewayPlugin",
+        "saleor.payment.gateways.dummy_credit_card.plugin.DummyCreditCardGatewayPlugin",
+    ]
+    return settings
+
+
+@pytest.fixture
+def sample_gateway(settings):
+    settings.PLUGINS += [
+        "saleor.plugins.tests.sample_plugins.ActiveDummyPaymentGateway"
+    ]
+    return settings
