+++ saleor/plugins/tests/fixtures/plugin_configuration.py
@@ -2,11 +2,9 @@ import copy
 
 import pytest
 
-from ..base_plugin import ConfigurationTypeField
-from ..manager import PluginsManager
-from ..models import PluginConfiguration
-from .sample_plugins import (
-    ALL_PLUGINS,
+from ...base_plugin import ConfigurationTypeField
+from ...models import PluginConfiguration
+from ..sample_plugins import (
     ChannelPluginSample,
     PluginInactive,
     PluginSample,
@@ -91,14 +89,3 @@ def new_config():
 @pytest.fixture
 def new_config_structure():
     return {"type": ConfigurationTypeField.STRING, "help_text": "foo", "label": "foo"}
-
-
-@pytest.fixture
-def plugins_manager():
-    return PluginsManager(plugins=[])
-
-
-@pytest.fixture
-def all_plugins_manager():
-    plugins_as_module_paths = [p.__module__ + "." + p.__name__ for p in ALL_PLUGINS]
-    return PluginsManager(plugins=plugins_as_module_paths)
